Change style / Renamed utils to utils_oocl / Add option for DAMCO
