+++ ndd-app/summary/views/summary_oocl_invoice.py
@@ -13,7 +13,7 @@ from django.views.generic import TemplateView
 
 from ndd.settings import STATICFILES_DIRS
 from booking.models import Booking
-from .utils import StyleXls
+from .utils_oocl import StyleXls
 from agent_transport.models import AgentTransport
 from customer.models import Principal, Shipper
 from ..models import Year, FormDetail, CustomerCustom, SummaryWeek, SummaryCustomer, Invoice, InvoiceDetail
