+++ ndd-app/summary/views/utils_oocl.py
