+++ ndd-app/templates/summary/summary_year_details_page.html
@@ -33,7 +33,7 @@
                         <td v-for="i in 12" :class="{'bg-secondary': ! year_detail.total[i-1]}" style="width:120px;">{{ year_detail.total[i-1]|currencyCommas }}</td>
                         <td style="width:120px;">{{ total_customer[index]|currencyCommas }}</td>
 
-                        <td v-if="year_detail.cusotomer_total" class="text-center">{{ year_detail.cusotomer_total|currencyCommas }}</td>
+                        <td v-if="year_detail.cusotomer_total" class="text-center"><b>{{ year_detail.cusotomer_total|currencyCommas }}</b></td>
                     </tr>
 
                     <tr class="text-right" >
